#1844 cut sql course druid error
